fix: split test drilling.spec
JIRA: QA-19249
